avoid a possible race condition when releasing h2 connections
It is possible that when two threads try to release streams of the same
h2 connection, one thread would register the connection in the in_use_pool
while the other register it in the idle_pool.

The solution is to put a lock so that the connection can only appear in
one pool at a time.